+++ android-app/app/src/main/java/com/sonnebtb/wowguildmanager/guildinteraction/calendar/CalendarAdapter.kt
@@ -8,10 +8,15 @@ import androidx.recyclerview.widget.RecyclerView
 import com.google.firebase.firestore.*
 import com.sonnebtb.wowguildmanager.Constants
 import com.sonnebtb.wowguildmanager.R
-import com.sonnebtb.wowguildmanager.guildinteraction.announcements.Announcement
+import com.sonnebtb.wowguildmanager.guildinteraction.FirebaseDeleteDelegate
 import com.sonnebtb.wowguildmanager.responses.Guild
 
-class CalendarAdapter(var context: Context?, var ref: CollectionReference, var guild: Guild) :
+class CalendarAdapter(
+    var context: Context?,
+    var ref: CollectionReference,
+    var guild: Guild,
+    var deleteDelegate: FirebaseDeleteDelegate
+) :
     RecyclerView.Adapter<CalendarViewHolder>() {
 
     var calendarEvents: ArrayList<CalendarEvent> = ArrayList()
@@ -58,9 +63,14 @@ class CalendarAdapter(var context: Context?, var ref: CollectionReference, var g
 
     override fun onBindViewHolder(holder: CalendarViewHolder, position: Int) {
         holder.bind(calendarEvents[position])
+
     }
 
     fun remove(position: Int){
-        ref.document(calendarEvents[position].id).delete()
+        if(deleteDelegate.userIsCreator(calendarEvents[position].creator!!)) {
+            ref.document(calendarEvents[position].id).delete()
+        } else {
+            Log.e(Constants.TAG, "User is not creator")
+        }
     }
 }
